% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_download.R
\name{download_display}
\alias{download_display}
\title{Download display full-load data}
\usage{
download_display(
  format = c("csv", "json"),
  destination_path = ".",
  unzip = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{format}{File format: \code{"csv"} or \code{"json"}.}

\item{destination_path}{Directory where ZIP is saved.}

\item{unzip}{Whether to extract ZIP contents into a sibling directory.}

\item{overwrite}{Whether to replace an existing ZIP.}
}
\description{
Download display full-load data
}
