% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_download.R
\name{download_info}
\alias{download_info}
\title{Fetch full-load download metadata}
\usage{
download_info(type = .epb_download_types, format = c("csv", "json"))
}
\arguments{
\item{type}{Dataset type: \code{"domestic"}, \code{"non_domestic"}, \code{"display"},
\code{"non_domestic_recommendation"}, or \code{"display_recommendation"}.}

\item{format}{File format: \code{"csv"} or \code{"json"}.}
}
\value{
A one-row tibble containing file size and last-updated timestamp.
}
\description{
Fetch full-load download metadata
}
