% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_download.R
\name{download_non_domestic_recommendations_json}
\alias{download_non_domestic_recommendations_json}
\title{Download non-domestic recommendation data}
\usage{
download_non_domestic_recommendations_json(
  destination_path = ".",
  unzip = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{destination_path}{Directory where ZIP is saved.}

\item{unzip}{Whether to extract ZIP contents into a sibling directory.}

\item{overwrite}{Whether to replace an existing ZIP.}
}
\description{
Download non-domestic recommendation data
}
