% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energizer-package.R
\docType{package}
\name{energizer-package}
\alias{energizer}
\alias{energizer-package}
\title{energizer: Client for the Energy Performance of Buildings Data API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R client for the UK government's Get energy performance of buildings data API. Provides bearer-authenticated access to Energy Performance Certificate (EPC) and Display Energy Certificate (DEC) data, including certificate retrieval, domestic, non-domestic and display searches, change events, code tables and monthly full-load downloads.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielegiraldo/energizer}
  \item \url{https://gggiraldo.github.io/enrgz/}
  \item Report bugs at \url{https://github.com/gabrielegiraldo/energizer/issues}
}

}
\author{
\strong{Maintainer}: Gabriele Giraldo \email{giraldo.gabriele@outlook.it}

}
\keyword{internal}
