% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_download.R
\name{epb_download_display_recommendations_json}
\alias{epb_download_display_recommendations_json}
\title{Download display recommendation data}
\usage{
epb_download_display_recommendations_json(
  destination_path = ".",
  unzip = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{destination_path}{Directory where ZIP is saved.}

\item{unzip}{Whether to extract ZIP contents into a sibling directory.}

\item{overwrite}{Whether to replace an existing ZIP.}
}
\description{
Download display recommendation data
}
