% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_codes.R
\name{epb_get_codes}
\alias{epb_get_codes}
\title{Fetch available EPC code tables}
\usage{
epb_get_codes()
}
\value{
A tibble with one \code{code} column.
}
\description{
Fetch available EPC code tables
}
\examples{
\dontrun{
epb_get_codes()
}
}
