% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epb_deltas.R
\name{get_deltas}
\alias{get_deltas}
\title{Fetch changed energy certificates}
\usage{
get_deltas(date_start, date_end = date_start)
}
\arguments{
\item{date_start}{Start date in \code{YYYY-MM-DD} format.}

\item{date_end}{End date in \code{YYYY-MM-DD} format. Defaults to \code{date_start}.}
}
\value{
A tibble containing certificate number, event type, and timestamp.
}
\description{
Retrieves certificates recorded as removed or having an updated UPRN during
a date range.
}
\examples{
\dontrun{
get_deltas("2025-01-01", "2025-01-31")
}
}
