% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energizer-package.R
\docType{package}
\name{energizer-package}
\alias{energizer}
\alias{energizer-package}
\title{energizer: A Client For Accessing The Open Data Communities EPC API}
\description{
An R client for the UK's Open Data Communities API, providing programmatic access to Energy Performance Certificate (EPC) and Display Energy Certificate (DEC) data for domestic, non-domestic, and public buildings. Includes functions for searching, retrieving certificates and recommendations, and downloading bulk data files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gggiraldo/energizer}
  \item Report bugs at \url{https://github.com/gggiraldo/energizer/issues}
}

}
\author{
\strong{Maintainer}: Gabriele Giraldo \email{giraldo.gabriele@outlook.it}

}
\keyword{internal}
