% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_defunct.R
\name{odc-defunct}
\alias{odc-defunct}
\alias{odc_set_key}
\alias{odc_get_data}
\alias{odc_search_data}
\alias{odc_bulk_download}
\alias{odc_get_file_list}
\alias{odc_get_file}
\alias{odc_get_meta}
\alias{odc_get_schema}
\title{Defunct Open Data Communities API functions}
\usage{
odc_set_key(...)

odc_get_data(...)

odc_search_data(...)

odc_bulk_download(...)

odc_get_file_list(...)

odc_get_file(...)

odc_get_meta(...)

odc_get_schema(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
These functions always throw a defunct-function error.
}
\description{
Open Data Communities API and Basic Authentication workflow were replaced by
Get energy performance of buildings data API. Use corresponding public API
function.
}
